[GEWISDB] Rename `Reckoning` sub-decision to `Statement`
