+++ module/Decision/src/Model/SubDecision/Financial/Statement.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Decision\Model\SubDecision;
+namespace Decision\Model\SubDecision\Financial;
 
 use Doctrine\ORM\Mapping\Entity;
 
 #[Entity]
-class Reckoning extends Budget
+class Statement extends Budget
 {
 }
